+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 * Remove deprecated `Kafka.spec.topicOperator` classes and deployment logic
 * Use Java 11 as the Java runtime
+* Upgrade to Cruise Control version 2.0.108
 
 ### Deprecations and removals